+++ packages/flutter_riverpod/lib/src/consumer.dart
@@ -10,7 +10,7 @@ abstract class WidgetRef {
   ///
   /// See also:
   ///
-  /// - [ProviderBase.select], which allows a widget to filter rebuilds by
+  /// - [ProviderListenable.select], which allows a widget to filter rebuilds by
   ///   observing only the properties.
   /// - [listen], to react to changes on a provider, such as for showing modals.
   T watch<T>(ProviderListenable<T> provider);