+++ packages/riverpod/lib/src/framework/container.dart
@@ -154,7 +154,7 @@ var _debugVerifyDependenciesAreRespectedEnabled = true;
 /// (outside of testing), as it is implicitly created for you by `ProviderScope`.
 /// {@endtemplate}
 @sealed
-class ProviderContainer {
+class ProviderContainer implements Node {
   /// {@macro riverpod.providercontainer}
   ProviderContainer({
     ProviderContainer? parent,
@@ -300,6 +300,23 @@ class ProviderContainer {
     return element.readSelf();
   }
 
+  @override
+  ProviderSubscription<State> _createSubscription<State>(
+    ProviderElementBase<State> element, {
+    required void Function(State? previous, State next) listener,
+    required void Function(Object error, StackTrace stackTrace) onError,
+  }) {
+    final sub = _ProviderSubscription<State>._(
+      element,
+      listener,
+      onError: onError,
+    );
+
+    element._listeners.add(sub);
+
+    return sub;
+  }
+
   /// Subscribe to this provider.
   ///
   /// See also:
@@ -308,6 +325,7 @@ class ProviderContainer {
   ///   closing the subscription.
   /// - [Ref.watch], which is an easier way for providers to listen
   ///   to another provider.
+  @override
   ProviderSubscription<State> listen<State>(
     ProviderListenable<State> provider,
     void Function(State? previous, State next) listener, {
@@ -315,19 +333,8 @@ class ProviderContainer {
     void Function(Object error, StackTrace stackTrace)? onError,
   }) {
     // TODO test always flushed provider
-    if (provider is _ProviderSelector<Object?, State>) {
-      return provider.listen(
-        this,
-        listener,
-        fireImmediately: fireImmediately,
-        onError: onError,
-      );
-    }
-
-    final element = readProviderElement(provider as ProviderBase<State>);
-
-    return element.addListener(
-      provider,
+    return provider.addListener(
+      this,
       listener,
       fireImmediately: fireImmediately,
       onError: onError,
@@ -442,12 +449,7 @@ class ProviderContainer {
     );
   }
 
-  /// Reads the state of a provider, potentially creating it in the process.
-  ///
-  /// It may throw if the provider requested threw when it was built.
-  ///
-  /// Do not use this in production code. This is exposed only for testing
-  /// and devtools, to be able to test if a provider has listeners or similar.
+  @override
   ProviderElementBase<State> readProviderElement<State>(
     ProviderBase<State> provider,
   ) {