+++ packages/riverpod/test/framework/provider_element_test.dart
@@ -68,7 +68,7 @@ void main() {
 
       expect(
         element.readSelf,
-        throwsA(isA<ProviderException>()),
+        throwsUnimplementedError,
       );
     });
   });